+++ src/components/Survivors.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import {useEffect} from "react";
 import {CoinScore} from "../../apps-script/shared-lib/types";
-import {Box, Button, IconButton, List, ListItem, ListItemAvatar, ListItemText, Stack} from "@mui/material";
+import {Alert, Box, Button, IconButton, List, ListItem, ListItemAvatar, ListItemText, Stack} from "@mui/material";
 import {Refresh} from "@mui/icons-material";
 
 export function Survivors() {
@@ -20,6 +20,10 @@ export function Survivors() {
   return (
     <Box sx={{justifyContent: 'center', display: 'flex'}}>
       <Stack spacing={2}>
+        <Alert sx={{width: 300}} severity={"info"}>
+          Score represents how many times a currency showed a price growth within
+          last {CoinScore.PRICES_MAX_CAP} measures, while 99% of the market was not moving up.
+        </Alert>
         {!!survivors.length &&
           <List sx={{padding: 0, width: 300}}>
             {survivors.map((rJson, i) => {